moved boardgames routes into dedicated file
